Add files via upload
The basic data set statistics consists of single experiments on the application of the ExMas algorithm to Uber data for the first half of 2016 for New York. The number of experiments is equal to the number of periods of half an hour, starting from 01/01/2016 00:00:00, ending on 06/30/2016 23:59:59 The discount level for ride-pooling is 0.32. The file presents the KPI obtained as a result of the experiments.